Build a custom pipe
A custome pipe to replace a character with spaces

Create the pipe component in the shared folder.
implement PipeTransform from @angular/core.

Use a Pipe decorator to provide the name that the custom pipe will be called by.

Implement the methods in the PipeTransform interface, tranform method.

Use the custom pipe after declaring it in the AppModule